Make use of new SLL capabilities
This allows to run the "runMods" task again without hell breaking loose.
The next stop would be AW negation, but I might be able to solve the
issue through other ways. Or perhaps this isn't an issue at all at
which point debugging support is "done" - even if I am not truely pleased
by how it is internally implemented.