Allow to include jar contents in jar without JIJ'ing the jar
For the uninitiated: JIJ = jar-in-jar

The former recommended solution of using the jar's sources directly
had the minor issue of being not really compatible with the `mrjar`
(multi-release-jar) plugin due to it also using directives such as
`into`, which would not get accounted for.

This solution - albeit being boilerplately - was identified to be
the best solution for the problem and should in theory allow
interoperability with any third-party plugin that modifies the jar.
It should also support the shadowJar plugin - but I have not tested
it yet.