Add experimental softmap support
Softmaps are rather experimental, so it is possible
that we move to more traditional static mappings formats
in the future.

Perhaps using an Enigma + Matcher + TinyV2 (using mapping-io)
toolchain combo isn't all too bad? Especially if snod
stops working on this game such a combdo would work well
(especially since the 'magic' Matcher component would be
superflouous). Alternatively we could stick on 5.0.2 forever,
but what is the point in that, my dear readers?