Allow use of supplementary mappings
A rather experimental feature, but it will probably be kept around.

What will probably not be kept around is tiny-remapper.
Performance-wise this usage of tiny-remapper is beyond dangerous,
though we will likely remove usage of TR altogether.
This means that we need to write a replacement for the mixin
remapping extension, but that should be in the realms of possibility.
Otherwise, most other things needed for a remapper are already
written across the stianloader toolchain (sl-deobf, softmap and
micromixin-transformer are the projects that primarily come to mind
here). Making a ClassNode-based remapper would have massive performance
boons even outside of supplementary mappings as I believe some
of our mixin extensions did the conversion to ClassNode anyways
for logistical reasons.