Do not transform jars before copying them over
I'm not exactly sure what I went for; I assume it was so that
obfuscated jars could be run in the development environment,
but that wasn't what it did in practice, so removing it is
safe to do (people shouldn't be using remapped/obfuscated
jars in the development environment anyways).

The breaking change that prompted me to remove this piece of code
was the fact that the ClassWriter wasn't correctly configured
as a result from us moving from TR to our own remapping framework.